+++ projects/packages/my-jetpack/src/products/class-videopress.php
@@ -156,4 +156,14 @@ class Videopress extends Hybrid_Product {
 			return admin_url( 'admin.php?page=jetpack#/settings?term=videopress' );
 		}
 	}
+
+	/**
+	 * Checks whether the current plan (or purchases) of the site already supports the product
+	 *
+	 * @return boolean
+	 */
+	public static function has_required_plan() {
+		// TODO: import and perform a proper check with Current_Plan. See #33410.
+		return true;
+	}
 }